Refactor: 여러 컴포넌트 및 파일 타입스크립트 마이그레이션
Refactor: 상수 타입스크립트 마이그레이션

Refactor: useModal 컴포넌트 타입스크립트 마이그레이션

Refactor: ReceivingPackage 컴포넌트 타입스크립트 마이그레이션

Refactor: Nav 컴포넌트 타입스크립트 마이그레이션

Refactor: MyPackages 컴포넌트 타입스크립트 마이그레이션

Refactor: InfoModal 컴포넌트 타입스크립트 마이그레이션

Refactor: Introduction 컴포넌트 타입스크립트 마이그레이션

Refactor: Home 컴포넌트 타입스크립트 마이그레이션

Refactor: CreatingPackage 컴포넌트 타입스크립트 마이그레이션

Refactor: main, App 컴포넌트 타입스크립트 마이그레이션

Refactor:  deleteFile 타입스크립트 마이그레이션 및 호출

+++ src/renderer/Components/App.tsx
@@ -46,7 +46,7 @@ function App() {
         setClientStatus({ isLogin: true });
       } catch (error) {
         console.error("앱초기진입 로그인에러: ", error);
-        if (error.name !== "TokenExpiredError") {
+        if (error instanceof Error && error.name !== "TokenExpiredError") {
           window.localStorage.clear();
           setClientStatus({ isLogin: false });
           return;
